+++ src/domain/swaps/contracts/decoders/gp-v2-decoder.helper.spec.ts
@@ -1,7 +1,7 @@
 import { Hex } from 'viem';
 import { faker } from '@faker-js/faker';
 import { GPv2Decoder } from '@/domain/swaps/contracts/decoders/gp-v2-decoder.helper';
-import { setPreSignatureEncoder } from '@/domain/swaps/contracts/__tests__/encoders/gp-v2-settlement-encoder.builder';
+import { setPreSignatureEncoder } from '@/domain/swaps/contracts/__tests__/encoders/gp-v2-encoder.builder';
 import { ILoggingService } from '@/logging/logging.interface';
 
 const loggingService = {
