+++ src/domain/swaps/contracts/decoders/gp-v2-decoder.helper.ts
@@ -21,7 +21,7 @@ export class GPv2Decoder extends AbiDecoder<typeof abi> {
    * @param data - the transaction data for the setPreSignature call
    * @returns {`0x${string}`} the order UID or null if the data does not represent a setPreSignature transaction
    */
-  getOrderUid(data: `0x${string}`): `0x${string}` | null {
+  getOrderUidFromSetPreSignature(data: `0x${string}`): `0x${string}` | null {
     try {
       if (!this.helpers.isSetPreSignature(data)) return null;
       const { args } = this.decodeFunctionData({ data });
